redid function-class modification
